refactor(quant): pre-extract quantized data types to foundation (Slice D)
Move the quantized DATA types (QuantizedVector + QuantEngineQuantizationMetadata +
the QuantizationMetadata alias + StorageQuantizedData) down from the modality kernel
to the foundation crate `proximadb-quantization-model` (alongside the
UnifiedQuantizationLevel cluster from #831), and redirect src/storage's references.

These types are foundation-pure (Vec<u8> + UnifiedQuantizationLevel [foundation, #831]
+ primitive metadata: codebook_id/scale/offset/norm + String + usize) — they had been
canonically in the kernel but only depended on foundation types. Moved verbatim;
re-exported from the kernel (`pub use`) so its public API + internal construction are
unchanged. Removed the now-unused `serde` import in quantization_engine.rs.

Why this matters beyond the ratchet: it makes the upcoming StorageQuantizationEnginePort
(D2b) FACADE-FREE — its return types (Vec<StorageQuantizedData>, &QuantizedVector) are
now foundation types, so the port needs no DTO/adaptor layer (unlike the assumption in
the Slice D plan). This is the incremental start of the factory-as-injection-boundary
effort.

- crates/foundation/proximadb-quantization-model: + the 4 data types.
- crates/modalities/proximadb-quantization-kernel: + model dep; defs -> `pub use` re-export.
- src/storage (progressive_search, raptor/common, viper/pipeline): 8 data-type refs ->
  proximadb_quantization_model::* (Mechanism A path rewrite).

Verified: cargo check --lib + clippy --lib clean; layering green (ceiling 205 -> 197);
storage up-edge ratchet 205 -> 197 (-8). Behavior-identical type move (recall-neutral).
No dependency on in-flight #596/#605 (verified: no file overlap).
